+++ src/components/lleidahacker/LleidaHackerHome/LleidaHackerHome.js
@@ -1,0 +1,10 @@
+import React from "react";
+
+
+const LleidaHackerHome = () => {
+  return (
+    <div>LleidaHackerHome</div>
+  )
+}
+
+export default LleidaHackerHome
